requirements.txt
streamlit
PyPDF2
langchain
langchain-google-genai
langchain-community
langchain-classic
google-generativeai
faiss-cpu
python-dotenv